Code cleanup!

+++ include/types.h
@@ -11,7 +11,7 @@ typedef unsigned long long ulonglong;
 
 typedef signed char s8;
 typedef signed short s16;
-typedef signed long s32;
+typedef signed int s32;
 typedef signed long long s64;
 typedef unsigned char u8;
 typedef unsigned short u16;
